+++ src/main/kotlin/com/menuplanner/server/menu/IngredientRepository.kt
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface IngredientRepository : JpaRepository<IngredientRecord, Long> {
-    fun findDistinctById(id: Long): List<IngredientRecord>
+interface IngredientRepository : JpaRepository<IngredientRecord, Int> {
+    fun findDistinctById(id: Int): List<IngredientRecord>
 }
